Refactor tests into colocated __tests__ directories
