+++ src/components/atoms/__tests__/LoadingSpinner.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 
-import LoadingSpinner from "./LoadingSpinner";
+import LoadingSpinner from "../LoadingSpinner";
 
 describe("LoadingSpinner", () => {
   it("renders with default medium size classes", () => {
