+++ src/components/atoms/__tests__/Logo.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import Logo from "./Logo";
+import Logo from "../Logo";
 
 describe("Logo", () => {
     it("renders the image logo", () => {
