+++ src/components/organisms/__tests__/Footer.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { MemoryRouter } from "react-router-dom";
 
-import Footer from "./Footer";
+import Footer from "../Footer";
 
 const scrollToElement = vi.fn();
